Upgrade Cobalt eqn, update Nagy, add fitting
Swapped Lambda_co equation for a mroe accurate one, updated the opacity "Nagy" function to better match the early data and not be undefined for times < t=3.4 due to a square root (Need to replace that entire function in future). Made a start on fitting data. Also added a little data viewer as a UI test.